chore: schema organization
