+++ app/tasks/actions.ts
@@ -1,6 +1,6 @@
 'use server';
 import { db } from '@/db';
-import { tasks } from '@/db/schema';
+import { tasks } from '@/db/schema/tasks';
 import { eq } from 'drizzle-orm';
 import { revalidatePath } from 'next/cache';
